refactor(test): inline mod tests を module/tests.rs へ分離 (#127)
diff coverage gate が #[cfg(test)] 内の防御アーム (panic!/unreachable! 等、
test 成功時に構造的に実行されない行) を変更行として数え uncovered 扱いする
blind spot を解消する。

- inline mod tests を持つ全 19 file を module/tests.rs へ分離
- ci.yml の llvm-cov に --ignore-filename-regex '/tests\.rs$' を追加し、
  分離した tests.rs を lcov から除外
- bin/eval_harness.rs は crate root のため #[path] で明示

Closes #127
